components/langfuse-bootstrap/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "fabric-langfuse-bootstrap"
description = "SingleAxis Fabric — curated-Langfuse init Job (Layer 1)."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.12"
authors = [{ name = "AI5Labs Research OPC Private Limited", email = "fabric@singleaxis.ai" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
  "httpx>=0.27,<1.0",
  "pydantic>=2.8,<3.0",
  "pyyaml>=6.0,<7.0",
  "typer>=0.12,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "respx>=0.21",
  "ruff>=0.6.9",
  "mypy>=1.11.2",
  "types-PyYAML>=6.0",
]

[project.scripts]
fabric-langfuse-bootstrap = "fabric_langfuse_bootstrap.__main__:main"

[tool.hatch.version]
path = "src/fabric_langfuse_bootstrap/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/fabric_langfuse_bootstrap"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
  "-ra",
  "--strict-config",
  "--strict-markers",
  "--cov=fabric_langfuse_bootstrap",
  "--cov-report=term-missing",
  "--cov-fail-under=80",
]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["fabric_langfuse_bootstrap"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "S", "C4", "PIE", "RET", "SIM", "PL", "RUF"]
ignore = ["PLR0913"]

[tool.ruff.lint.per-file-ignores]
# S105/S106 flag hardcoded "test" credentials in fixtures; these aren't
# real secrets, they're just the strings tests use to exercise the
# client auth path.
"tests/**" = ["S101", "S105", "S106", "PLR2004"]
# typer.Option(...) in argument defaults is the library's ergonomic;
# suppressing B008 here keeps the CLI entrypoint idiomatic.
"src/fabric_langfuse_bootstrap/__main__.py" = ["B008"]
# HTTP status codes 200/404 are well-understood; naming them is worse
# than the number. Late imports break an import cycle documented in
# the module docstring.
"src/fabric_langfuse_bootstrap/client.py" = ["PLR2004", "PLC0415"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
